Move references helpers to TreeUtils
https://github.com/digital-asset/daml/pull/8912#discussion_r580396852

Also, use `scalaz.std.iterable._` in `TreeUtils`.
